refactor: improve fetch api by using service
