+++ src/main.tsx
@@ -6,6 +6,7 @@ import { UserFlex } from './pages/userflex/UserFlex';
 import { UserGrid } from './pages/usergrid/UserGrid';
 import { NotFound } from './pages/NotFound';
 import NavBar from './components/navbar/NavBar';
+import { getUsers, mapAndSaveUsers } from '../src/services/userService';
 
 type User = {
   id: number
@@ -13,48 +14,15 @@ type User = {
   description: string
 }
 
-type ApiUser = {
-  id: number
-  name: string
-  username: string
-  email: string
-  address: {
-      street: string
-      suite: string
-      city: string
-      zipcode: string
-      geo: {
-          lat: string
-          lng: string
-      }
-  },
-  phone: string
-  website: string
-  company: {
-      name: string
-      catchPhrase: string
-      bs: string
-  }
-}
-
 export const UserContext = React.createContext<User[]>([]);
 
 export const App = (): JSX.Element => {
   const [users, setUsers] = useState<User[]>([]);
 
-  const mapAndSaveUsers = (users: ApiUser[]): void => {
-      const mappedUser = users.map(user => ({
-          id: user.id, 
-          name: user.name, 
-          description: user.company.catchPhrase
-      }))
-      setUsers(mappedUser);
-  }
-
   useEffect(() => {
-      fetch('https://jsonplaceholder.typicode.com/users')
-      .then(response => response.json())
-      .then(mapAndSaveUsers)
+      getUsers().then((users) => {
+        setUsers(mapAndSaveUsers(users));
+      })
   }, []);
 
   return (
